pyproject.toml
[tool.poetry]
name = "emr-api"
version = "0.1.0"
description = ""
authors = ["Lana <lan2828@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
drf-yasg = "^1.21.8"
pytest-cov = "^6.0.0"
coverage = "^7.6.10"
isort = "^5.13.2"
flake8 = "^7.1.1"
black = "^24.10.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
